#6811 - Created a separate service for travel entries list
