+++ sormas-backend/src/main/java/de/symeda/sormas/backend/travelentry/TravelEntryService.java
@@ -24,7 +24,6 @@ import de.symeda.sormas.api.EntityRelevanceStatus;
 import de.symeda.sormas.api.task.TaskCriteria;
 import de.symeda.sormas.api.travelentry.TravelEntryCriteria;
 import de.symeda.sormas.api.travelentry.TravelEntryIndexDto;
-import de.symeda.sormas.api.travelentry.TravelEntryListEntryDto;
 import de.symeda.sormas.api.travelentry.TravelEntryReferenceDto;
 import de.symeda.sormas.api.user.UserRight;
 import de.symeda.sormas.api.utils.DataHelper;
@@ -41,7 +40,6 @@ import de.symeda.sormas.backend.person.Person;
 import de.symeda.sormas.backend.task.Task;
 import de.symeda.sormas.backend.task.TaskService;
 import de.symeda.sormas.backend.travelentry.transformers.TravelEntryIndexDtoResultTransformer;
-import de.symeda.sormas.backend.travelentry.transformers.TravelEntryListEntryDtoResultTransformer;
 import de.symeda.sormas.backend.user.User;
 import de.symeda.sormas.backend.user.UserService;
 import de.symeda.sormas.backend.util.JurisdictionHelper;
@@ -159,39 +157,6 @@ public class TravelEntryService extends AbstractCoreAdoService<TravelEntry> {
 
 		cq.select(cb.countDistinct(travelEntry));
 		return em.createQuery(cq).getSingleResult();
-	}
-
-	public List<TravelEntryListEntryDto> getEntriesList(Long personId, String caseUuid, Integer first, Integer max) {
-		final CriteriaBuilder cb = em.getCriteriaBuilder();
-		final CriteriaQuery<Object[]> cq = cb.createQuery(Object[].class);
-		final Root<TravelEntry> travelEntry = cq.from(TravelEntry.class);
-
-		final TravelEntryQueryContext travelEntryQueryContext = new TravelEntryQueryContext(cb, cq, travelEntry);
-
-		final TravelEntryJoins<TravelEntry> joins = (TravelEntryJoins<TravelEntry>) travelEntryQueryContext.getJoins();
-		final Join<TravelEntry, PointOfEntry> pointOfEntry = joins.getPointOfEntry();
-
-		cq.multiselect(
-			travelEntry.get(TravelEntry.UUID),
-			travelEntry.get(TravelEntry.REPORT_DATE),
-			travelEntry.get(TravelEntry.DISEASE),
-			pointOfEntry.get(PointOfEntry.NAME),
-			JurisdictionHelper.booleanSelector(cb, inJurisdictionOrOwned(travelEntryQueryContext)),
-			travelEntry.get(TravelEntry.CHANGE_DATE));
-
-		final Predicate criteriaFilter = buildListEntryCriteriaFilter(personId, caseUuid, travelEntryQueryContext);
-
-		if (criteriaFilter != null) {
-			cq.where(criteriaFilter);
-		}
-
-		cq.orderBy(cb.desc(travelEntry.get(TravelEntry.CHANGE_DATE)));
-
-		cq.distinct(true);
-
-		return createQuery(cq, first, max).unwrap(org.hibernate.query.Query.class)
-			.setResultTransformer(new TravelEntryListEntryDtoResultTransformer())
-			.getResultList();
 	}
 
 	public boolean inJurisdictionOrOwned(TravelEntry travelEntry) {
@@ -384,28 +349,6 @@ public class TravelEntryService extends AbstractCoreAdoService<TravelEntry> {
 		}
 
 		filter = CriteriaBuilderHelper.and(cb, filter, createDefaultFilter(cb, from));
-		return filter;
-	}
-
-	private Predicate buildListEntryCriteriaFilter(Long personId, String caseUuid, TravelEntryQueryContext travelEntryQueryContext) {
-
-		final TravelEntryJoins joins = (TravelEntryJoins) travelEntryQueryContext.getJoins();
-		final CriteriaBuilder cb = travelEntryQueryContext.getCriteriaBuilder();
-		final From<?, TravelEntry> from = travelEntryQueryContext.getRoot();
-		Join<TravelEntry, Case> resultingCase = joins.getResultingCase();
-
-		Predicate filter = createUserFilter(travelEntryQueryContext);
-
-		if (personId != null) {
-			filter = CriteriaBuilderHelper.and(cb, filter, cb.equal(from.get(TravelEntry.PERSON_ID), personId));
-		}
-
-		if (caseUuid != null) {
-			filter = CriteriaBuilderHelper.and(cb, filter, cb.equal(resultingCase.get(Case.UUID), caseUuid));
-		}
-
-		filter = CriteriaBuilderHelper.and(cb, filter, cb.isFalse(from.get(TravelEntry.DELETED)));
-
 		return filter;
 	}
 }
